【OS X】修复编译错误
(cherry picked from commit a98db728a31ad880becadb2183265c5e47263ece)

+++ source/Platform/Source/Adapter/OSX/T3DOSXDeviceInfo.mm
@@ -451,7 +451,7 @@ namespace Tiny3D
         return mCPUCores;
     }
 
-    uint32_t OSXDeviceInfo::getSystemRAM() const
+    uint64_t OSXDeviceInfo::getSystemRAM() const
     {
         return mSystemRAM;
     }
